Conflicts:
	.cproject
	.settings/org.eclipse.cdt.managedbuilder.core.prefs
	src/Alien.cpp
	src/Bullet.cpp
	src/Camera.cpp
	src/EndState.cpp
	src/Game.cpp
	src/Game.h
	src/GameObject.h
	src/InputManager.cpp
	src/InputManager.h
	src/Main.cpp
	src/Penguins.cpp
	src/Penguins.h
	src/Player.cpp
	src/Player.h
	src/Sprite.cpp
	src/Sprite.h
	src/StageState.cpp
	src/StageState.h
	src/State.cpp
	src/State.h
	src/Text.cpp
	src/Text.h
	src/TileMap.cpp
	src/TileMap.h
	src/TileSet.cpp
	src/TileSet.h
	src/TitleState.cpp
	src/TitleState.h

+++ src/TileSet.cpp
@@ -1,41 +1,34 @@
-/*
- * TileSet.cpp
- *
- *  Created on: 01/04/2015
- *      Author: Vitor
- */
-
 #include "TileSet.h"
 
-TileSet::TileSet(int aux_tileWidth, int aux_tileHeight, string file) : tileSet(file){
+TileSet::TileSet(int aux_tileWidth, int aux_tileHeight, string file) :
+		tileSet(file) {
 	tileWidth = aux_tileWidth;
 	tileHeight = aux_tileHeight;
-	if(!tileSet.IsOpen()){
-		cout<<"Erro ao abrir sprite do tile set!!" << endl;
+	if (!tileSet.IsOpen()) {
+		cout << "Erro ao abrir sprite do tile set!!" << endl;
 		exit(1);
 	}
-	rows = tileSet.GetHeight()/tileHeight;
-	cout<< "rows : " << rows << endl;
-	columns = tileSet.GetWidth()/tileWidth;
-	cout<< "columns : " << columns << endl;
+	rows = tileSet.GetHeight() / tileHeight;
+	cout << "rows : " << rows << endl;
+	columns = tileSet.GetWidth() / tileWidth;
+	cout << "columns : " << columns << endl;
 }
 
-
-void TileSet::Render(unsigned int index,float x,float y,float angle){
-	if(index > ((rows*columns) - 1)){
+void TileSet::Render(unsigned int index, float x, float y, float angle) {
+	if (index > (unsigned int) ((rows * columns) - 1)) {
 		cout << "Index de tileSet inválido!!  - " << index << endl;
 		exit(1);
 	}
-	int sp_x = index%columns;
-	int sp_y = floor(index/columns);
-	tileSet.SetClip(sp_x*tileWidth,sp_y*tileHeight,tileWidth,tileHeight);
-	tileSet.Render(x,y,angle);
+	int sp_x = index % columns;
+	int sp_y = floor(index / columns);
+	tileSet.SetClip(sp_x * tileWidth, sp_y * tileHeight, tileWidth, tileHeight);
+	tileSet.Render(x, y, angle);
 }
 
-int TileSet::GetTileWidth(){
+int TileSet::GetTileWidth() {
 	return tileWidth;
 }
 
-int TileSet::GetTileHeight(){
+int TileSet::GetTileHeight() {
 	return tileHeight;
 }
